+++ index.ts
@@ -1,7 +1,1 @@
-export {
-    TagInputModule,
-    TagInputComponent,
-    TagInputForm,
-    TagInputDropdown
-} from './dist/modules/ng2-tag-input.module';
-
+export * from './dist/modules/index';