+++ modules/components/index.ts
@@ -1,1 +1,5 @@
+export * from './dropdown';
+export * from './icon';
+export * from './tag-input-form';
+export * from './tag';
 export * from './tag-input';