+++ modules/components/tag-input-form/tag-input-form.component.ts
@@ -3,7 +3,6 @@ import {
     Input,
     Output,
     EventEmitter,
-    Renderer,
     ViewChild
 } from '@angular/core';
 
@@ -12,7 +11,8 @@ import {
     FormControl,
     Validators,
     ValidatorFn,
-    AbstractControl
+    AbstractControl,
+    AsyncValidatorFn
 } from '@angular/forms';
 
 @Component({
@@ -66,6 +66,13 @@ export class TagInputForm {
     @Input() public validators: ValidatorFn[] = [];
 
     /**
+     * @name asyncValidators
+     * @desc array of AsyncValidator that are used to validate the tag before it gets appended to the list
+     * @type {Array}
+     */
+    @Input() public asyncValidators: AsyncValidatorFn[] = [];
+
+    /**
      * @name inputId
      * @type {string}
      */
@@ -90,7 +97,7 @@ export class TagInputForm {
      * @type {string}
      */
     @Input() public tabindex: string = undefined;
-  
+
     /**
      * @name disabled
      */
@@ -126,12 +133,15 @@ export class TagInputForm {
      */
     public inputTextValue = '';
 
-    constructor(private renderer: Renderer) {}
+    constructor() {}
 
     public ngOnInit() {
         // creating form
         this.form = new FormGroup({
-            item: new FormControl('', Validators.compose(this.validators))
+            item: new FormControl('',
+                Validators.compose(this.validators),
+                Validators.composeAsync(this.asyncValidators)
+            )
         });
     }
 
@@ -176,14 +186,14 @@ export class TagInputForm {
      * @name focus
      */
     public focus(): void {
-        this.renderer.invokeElementMethod(this.input.nativeElement, 'focus');
+        this.input.nativeElement.focus();
     }
 
     /**
      * @name blur
      */
     public blur(): void {
-        this.renderer.invokeElementMethod(this.input.nativeElement, 'blur');
+        this.input.nativeElement.blur();
     }
 
 	/**
@@ -200,7 +210,7 @@ export class TagInputForm {
      */
     public destroy(): void {
         const input = this.input.nativeElement;
-        this.renderer.invokeElementMethod(input.parentElement, 'removeChild', [input]);
+        input.parentElement.removeChild(input);
     }
 
     /**