+++ modules/components/tag/index.ts
@@ -1,0 +1,2 @@
+export * from './tag-ripple.component';
+export * from './tag.component';