+++ modules/index.ts
@@ -1,14 +1,19 @@
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
+
 import { Ng2DropdownModule } from 'ng2-material-dropdown';
-import { DeleteIconComponent } from './components/icon';
-import { TagInputForm } from './components/tag-input-form';
-import { TagInputComponent } from './components';
-import { TagInputDropdown } from './components/dropdown/tag-input-dropdown.component';
-import { HighlightPipe } from './components/pipes/highlight.pipe';
-import { TagComponent } from './components/tag/tag.component';
-import { TagRipple } from './components/tag/tag-ripple.component';
+
+import {
+    TagInputComponent,
+    TagInputForm,
+    DeleteIconComponent,
+    TagInputDropdown,
+    TagComponent,
+    TagRipple
+} from './components';
+
+import { HighlightPipe } from './core';
 
 @NgModule({
     imports: [
@@ -38,11 +43,4 @@ import { TagRipple } from './components/tag/tag-ripple.component';
 })
 export class TagInputModule {}
 
-export {
-    TagInputComponent,
-    TagInputForm,
-    TagInputDropdown,
-    DeleteIconComponent,
-    TagComponent,
-    TagRipple
-}
+export * from './components';