+++ webpack.test.js
@@ -1,10 +1,26 @@
 const path = require('path');
+const ROOT = path.resolve(__dirname, '.');
+const root = path.join.bind(path, ROOT);
 
 module.exports = {
     devtool: 'cheap-module-source-map',
 
     resolve: {
-        extensions: ['.ts', '.js']
+        extensions: ['.ts', '.js'],
+        alias: {
+            'modules': root('modules'),
+            'core': root('modules/core'),
+            'components': root('modules/components'),
+            'styles': root('modules/core/styles'),
+            'pipes': root('modules/core/pipes'),
+            'helpers': root('modules/core/helpers'),
+
+            'tag-input': root('modules/components/tag-input'),
+            'dropdown': root('modules/components/dropdown'),
+            'icon': root('modules/components/icon'),
+            'tag': root('modules/components/tag'),
+            'tag-input-form': root('modules/components/tag-input-form'),
+        }
     },
 
     module: {